+++ src/index.ts
@@ -34,7 +34,7 @@ function processNoInput<T>(arg: T): T {
     setNoInput(true);
   }
   if (hasNoInput()) {
-    logger.info('[no-input] The script will not accept any input!');
+    logger.debug('[no-input] The script will not accept any input!');
   }
   return arg;
 }