+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - fix(publish): Fix publishing when resuming from a state file (#197)
+- ref(logger): Move some extra info to debug level (#198)
 
 ## 0.19.0